Add a task to prepare deployment artifacts
In our dapss we expect deployment artifacts to be placed in `artifacts`
directory in flat structure. The `hardhat-plugin` exports the artifacts
to a directory `deployments/<network>`. We need to copy the artifacts
before publishing a package.

The task we add here will handle that.

To execute it we will call e.g.:
```
npx hardhat prepare-artifacts --network goerli
```