Output error instead of throwing it when verification fails
If verification of some contract on Etherscan fails, we don't want to throw an
error (which stops verification process), but we just want to list the error in
log. This is because when we're deploying & verifying  multiple contracts we
don't want to stop deployment of next contracts after failure in verification of
a contract in the middle of the list.