Publish feature packages without creating a commit
When workflow gets triggered on a feature branch we don't want to create a
commit automatically, because such automatically created commits are not marked
as verified and with current settings of GH repo they block the PRs from
merging. For such use case we decided to publish the packages under `feature`
suffix (preid), without creating a commit.