Extrct deployment transaction from upgraded contract
After proxy contract is upgraded we had a problem with
extracting deployment transaction via
`contract.deploymentTransaction()` from ethers. It was happening
because OpenZeppelin's `upgradeProxy` was replacing deployment
tx on the proxy contract but the field was different than what ethers
function expected. Let's make it work by directly using the field
that `upgradeProxy` adds on the contact's object.